refactor: 폴더명 수정 및 일부 코드 리팩토링
